added small fixes
