+++ src/types/auth-user-data.ts
@@ -1,0 +1,10 @@
+export type AuthData = {
+  login: string;
+  password: string;
+}
+
+export type UserData = {
+  id: number;
+  email: string;
+  token: string;
+}
